==> code/01_data-collection.ipynb <==
# 01 - Data Collection (Reddit API)

# Import libraries

import requests
import pandas as pd
import time

# Reddit API URL for posts

base_url = 'https://api.pushshift.io/reddit/submission/search'

%time

# Create empty list for vegan posts

vegan_posts = []

params_until = {
    'subreddit': 'vegan',
    'limit': 1000,
    'until': 1679865168,    # posts collected through 3/26/2023, 4:12:48 PM CST
                            # (1679865168 seconds since 1/1/1970, 12 AM GMT) 
    'sort': 'created_utc',
    'order': 'desc'
}

num_posts = 0
total_time = 0

for _ in range(100):  # scrape for up to 100k posts
    start = time.time()
    res = requests.get(base_url, params_until)
    print(res.status_code) # view status, error codes
    try:
        posts = pd.DataFrame(res.json()['data'])
        posts = posts[['title', 'selftext', 'created_utc']] 
        vegan_posts.append(posts)
        params_until['until'] = posts['created_utc'].min()  # go back to earliest post in group
        num_posts += len(posts) # show number of posts collected
        end = time.time()
        time_to = end - start # amount of time to scrape
        total_time += time_to
    except:
        pass
    #build in SOME cleaning
    print(f"time to scrape: {time_to}; num posts: {num_posts}") # print status

%time

# same process as for vegan

plant_posts = []

params_until = {
    'subreddit': 'PlantBasedDiet',
    'limit': 1000,
    'until': 1679865168,
    'sort': 'created_utc',
    'order': 'desc'
}

num_posts = 0
total_time = 0

for _ in range(100):
    start = time.time()
    res = requests.get(base_url, params_until)
    print(res.status_code) 
    try:
        posts = pd.DataFrame(res.json()['data'])
        posts = posts[['title', 'selftext', 'created_utc']] 
        plant_posts.append(posts)
        params_until['until'] = posts['created_utc'].min()
        num_posts += len(posts) 
        end = time.time()
        time_to = end - start
        total_time += time_to
    except:
        pass
    print(f"time to scrape: {time_to}; num posts: {num_posts}") 

# concatenate and export to data folder

pd.concat(vegan_posts).to_csv('../data/vegan.csv',index=False)
pd.concat(plant_posts).to_csv('../data/plant.csv',index=False)
