==> code/03_pre-processing.ipynb <==
# 03 Pre-Processing

# Import libraries

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.compose import ColumnTransformer

import spacy

# Read clean data

vegan = pd.read_csv('../data/vegan_clean.csv')
plant = pd.read_csv('../data/plant_clean.csv')

# Set values for target

vegan['target'] = 0
plant['target'] = 1

# Concatenate posts

posts = pd.concat([vegan,plant], axis = 0)

# Check for nulls

posts.isnull().sum()

#  Drop nulls

posts.dropna(inplace=True)

# Select title and selftext as features

X = posts[['title','selftext']]
y = posts['target']

# Train-test split

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    stratify = y, 
                                                    random_state = 42)

# View shape of data 

X_train.shape, y_train.shape, X_test.shape, y_test.shape

# View distribution of target for baseline accuracy score

print(y_train.value_counts(normalize=True))
print(y_test.value_counts(normalize=True))

Both train and target have approximately 72% vegan posts, so this is the baseline accuracy score that our models will attempt to beat.

#  Create spacy object

nlp = spacy.load('en_core_web_md')

stopwords = nlp.Defaults.stop_words

#  Create spacy tokenizer
#  Tokenize will lemmatize and exclude lemmas with POS unrelated to post content

class spacy_tokenizer():
    def __init__(self):
        self
    def __call__(self,post):
        doc = nlp(post)
        return [token.lemma_ for token in doc if 
                         token.pos_.lower() not in ['aux','punct','cconj','det','space','conj','adp','pron','sym']]

#  TF-IDF Vectorizer for titles and selftext

tvec_t = TfidfVectorizer(stop_words = stopwords, 
                         max_features = 1000,
                         min_df = 3,
                         max_df = .95,
                         ngram_range = (1,1),
                         tokenizer = spacy_tokenizer())

tvec_s = TfidfVectorizer(stop_words = stopwords, 
                         max_features = 1000,
                         min_df = 3,
                         max_df = .95,
                         ngram_range = (1,1),
                         tokenizer = spacy_tokenizer())


We are using small values for max_features and ngram_range and not checking combinations through GridSearch because of the computational constraints of the tokenizer and later constraints that would arise from combining the vectorizer to a pipeline. 

ctx = ColumnTransformer([
    ('tvec_t',tvec_t,'title'),
    ('tvec_s',tvec_s,'selftext')
],remainder = 'passthrough',n_jobs = -1)

ctx.fit(X_train)

X_train = ctx.transform(X_train)

X_test = ctx.transform(X_test)

#  Check shapes (should have 1000 columns each)

X_train.shape,X_test.shape

#  Reassemble as dataframes

X_train = pd.DataFrame(X_train.toarray(), columns = ctx.get_feature_names_out())
X_test = pd.DataFrame(X_test.toarray(), columns = ctx.get_feature_names_out())

# Write vectorized data to data folder

X_train.to_csv('../data/X_train_tvec.csv',index=False)
X_test.to_csv('../data/X_test_tvec.csv',index=False)

# Write targets to data folder

y_train.to_csv('../data/y_train.csv',index=False)
y_test.to_csv('../data/y_test.csv',index=False)
